--- Corona-Warn-App/src/main/java/de/rki/coronawarnapp/bugreporting/debuglog/ui/share/DebugLogShareNavigationEvents.kt
@@ -1,5 +1,0 @@
-package de.rki.coronawarnapp.bugreporting.debuglog.ui.share
-
-sealed class DebugLogShareNavigationEvents {
-    object NavigateToMoreInformationFragment : DebugLogShareNavigationEvents()
-}
