+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/bugreporting/debuglog/ui/upload/DebugLogUploadFragment.kt
@@ -1,11 +1,11 @@
-package de.rki.coronawarnapp.bugreporting.debuglog.ui.share
+package de.rki.coronawarnapp.bugreporting.debuglog.ui.upload
 
 import android.os.Bundle
 import android.view.View
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import de.rki.coronawarnapp.R
-import de.rki.coronawarnapp.databinding.BugreportingDebuglogShareFragmentBinding
+import de.rki.coronawarnapp.databinding.BugreportingDebuglogUploadFragmentBinding
 import de.rki.coronawarnapp.util.di.AutoInject
 import de.rki.coronawarnapp.util.ui.observe2
 import de.rki.coronawarnapp.util.ui.popBackStack
@@ -14,11 +14,11 @@ import de.rki.coronawarnapp.util.viewmodel.CWAViewModelFactoryProvider
 import de.rki.coronawarnapp.util.viewmodel.cwaViewModels
 import javax.inject.Inject
 
-class DebugLogShareFragment : Fragment(R.layout.bugreporting_debuglog_share_fragment), AutoInject {
+class DebugLogUploadFragment : Fragment(R.layout.bugreporting_debuglog_upload_fragment), AutoInject {
 
     @Inject lateinit var viewModelFactory: CWAViewModelFactoryProvider.Factory
-    private val vm: DebugLogShareViewModel by cwaViewModels { viewModelFactory }
-    private val binding: BugreportingDebuglogShareFragmentBinding by viewBindingLazy()
+    private val vm: DebugLogUploadViewModel by cwaViewModels { viewModelFactory }
+    private val binding: BugreportingDebuglogUploadFragmentBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -38,7 +38,7 @@ class DebugLogShareFragment : Fragment(R.layout.bugreporting_debuglog_share_frag
         vm.routeToScreen.observe2(this) {
             when (it) {
 
-                DebugLogShareNavigationEvents.NavigateToMoreInformationFragment -> {
+                DebugLogUploadNavigationEvents.NavigateToMoreInformationFragment -> {
                     // TODO Add navigation to new screen
                 }
             }
