+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/bugreporting/debuglog/ui/upload/DebugLogUploadNavigationEvents.kt
@@ -1,0 +1,5 @@
+package de.rki.coronawarnapp.bugreporting.debuglog.ui.upload
+
+sealed class DebugLogUploadNavigationEvents {
+    object NavigateToMoreInformationFragment : DebugLogUploadNavigationEvents()
+}
