+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/bugreporting/debuglog/ui/upload/DebugLogUploadViewModel.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.bugreporting.debuglog.ui.share
+package de.rki.coronawarnapp.bugreporting.debuglog.ui.upload
 
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
@@ -7,15 +7,15 @@ import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModel
 import de.rki.coronawarnapp.util.viewmodel.SimpleCWAViewModelFactory
 
-class DebugLogShareViewModel @AssistedInject constructor(
+class DebugLogUploadViewModel @AssistedInject constructor(
     dispatcherProvider: DispatcherProvider,
 ) : CWAViewModel(dispatcherProvider = dispatcherProvider) {
-    val routeToScreen: SingleLiveEvent<DebugLogShareNavigationEvents> = SingleLiveEvent()
+    val routeToScreen: SingleLiveEvent<DebugLogUploadNavigationEvents> = SingleLiveEvent()
 
     fun onUploadLog() {
         // TODO Implement Uploading
     }
 
     @AssistedFactory
-    interface Factory : SimpleCWAViewModelFactory<DebugLogShareViewModel>
+    interface Factory : SimpleCWAViewModelFactory<DebugLogUploadViewModel>
 }
