+++ Corona-Warn-App/src/main/res/layout/bugreporting_debuglog_upload_fragment.xml
