+++ Corona-Warn-App/src/main/res/navigation/nav_graph.xml
@@ -510,8 +510,8 @@
             android:id="@+id/action_debuglogFragment_to_logUploadHistoryFragment"
             app:destination="@id/logUploadHistoryFragment" />
         <action
-            android:id="@+id/action_debuglogFragment_to_debugLogShareFragment"
-            app:destination="@id/debugLogShareFragment" />
+            android:id="@+id/action_debuglogFragment_to_debugLogUploadFragment"
+            app:destination="@id/debugLogUploadFragment" />
     </fragment>
 
     <!-- New Release -->
@@ -592,8 +592,8 @@
         android:label="LogUploadHistoryFragment"
         tools:layout="@layout/bugreporting_upload_history_fragment" />
     <fragment
-        android:id="@+id/debugLogShareFragment"
-        android:name="de.rki.coronawarnapp.bugreporting.debuglog.ui.share.DebugLogShareFragment"
-        android:label="DebugLogShareFragment"
-        tools:layout="@layout/bugreporting_debuglog_share_fragment"/>
+        android:id="@+id/debugLogUploadFragment"
+        android:name="de.rki.coronawarnapp.bugreporting.debuglog.ui.upload.DebugLogUploadFragment"
+        android:label="DebugLogUploadFragment"
+        tools:layout="@layout/bugreporting_debuglog_upload_fragment" />
 </navigation>
